Replacesd existing example with 2 new examples, each in their own subfolder
The existing example was not clear enought and has been replaced with 2 different integration approaches: using a singleton and using notifications.
