--- Podfile
@@ -1,12 +1,0 @@
-# Uncomment the next line to define a global platform for your project
-# platform :ios, '10.0'
-
-target 'TapResearchSwiftDemo' do
-  # Comment the next line if you don't want to use dynamic frameworks
-  use_frameworks!
-
-  # Pods for TapResearchSwiftDemo
-
-pod 'TapResearch', '2.5.8'
-
-end
