--- TapResearchSwiftDemo/AppDelegate.swift
@@ -1,24 +1,0 @@
-//
-//  AppDelegate.swift
-//  TapResearchSwiftDemo
-//
-//  Created by Jeroen Verbeek on 12/7/20.
-//
-
-import UIKit
-import TapResearchSDK
-
-@main
-class AppDelegate: UIResponder, UIApplicationDelegate {
-
-    func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?) -> Bool {
-        return true
-    }
-
-    func application(_ application: UIApplication, configurationForConnecting connectingSceneSession: UISceneSession, options: UIScene.ConnectionOptions) -> UISceneConfiguration {
-        return UISceneConfiguration(name: "Default Configuration", sessionRole: connectingSceneSession.role)
-    }
-    
-}
-
-
