+++ UsingNotifications/Podfile
@@ -1,0 +1,12 @@
+# Uncomment the next line to define a global platform for your project
+# platform :ios, '10.0'
+
+target 'TapResearchSwiftDemoNotifications' do
+  # Comment the next line if you don't want to use dynamic frameworks
+  use_frameworks!
+
+  # Pods for TapResearchSwiftDemoNotifications
+
+pod 'TapResearch', '2.5.13'
+
+end
