+++ UsingSingleton/TapResearchSwiftDemoSingleton.xcodeproj/project.xcworkspace/xcshareddata/WorkspaceSettings.xcsettings
@@ -1,0 +1,10 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
+<plist version="1.0">
+<dict>
+	<key>DisableBuildSystemDeprecationDiagnostic</key>
+	<true/>
+	<key>PreviewsEnabled</key>
+	<false/>
+</dict>
+</plist>
