+++ UsingSingleton/TapResearchSwiftDemoSingleton/AppDelegate.swift
@@ -1,0 +1,25 @@
+//
+//  AppDelegate.swift
+//  TapResearchSwiftDemoSingleton
+//
+//  Created by Jeroen Verbeek on 12/7/20.
+//
+
+import UIKit
+
+@main
+class AppDelegate: UIResponder, UIApplicationDelegate {
+
+    func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?) -> Bool {
+
+        TapHandler.shared.initSDK("<<YOUR_API_TOKEN>>")
+        return true
+    }
+
+    func application(_ application: UIApplication, configurationForConnecting connectingSceneSession: UISceneSession, options: UIScene.ConnectionOptions) -> UISceneConfiguration {
+        return UISceneConfiguration(name: "Default Configuration", sessionRole: connectingSceneSession.role)
+    }
+    
+}
+
+
